PATCH: Clean and fix user controller

+++ service/src/main/java/ar/edu/itba/paw/service/UserServiceImpl.java
@@ -6,12 +6,14 @@ import ar.edu.itba.paw.interfaces.services.ImageService;
 import ar.edu.itba.paw.interfaces.services.UserService;
 import ar.edu.itba.paw.model.Image;
 import ar.edu.itba.paw.model.User;
+import ar.edu.itba.paw.model.exceptions.ForbiddenUserException;
 import ar.edu.itba.paw.model.exceptions.UserNotFoundException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.Authentication;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.context.SecurityContextHolder;
@@ -69,14 +71,17 @@ public class UserServiceImpl implements UserService {
     @Transactional
     @Override
     public User updateImage(String username, byte[] data) {
-        User user = getUserByUsername(username).orElseThrow(() -> {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        if (!username.equals(auth.getName())) {
+            throw new ForbiddenUserException();
+        }
+
+        final User user = getUserByUsername(username).orElseThrow(() -> {
             LOGGER.error("Cannot update image for user {} because it does not exist", username);
             return new UserNotFoundException();
         });
 
-        Image image = null;
-        if (user.getImage() != null)
-            image = user.getImage();
+        final Image image = user.getImage();
 
         user.updateImage(data);
 
@@ -86,21 +91,25 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public User deleteImage(String username) {
-        User user = getUserByUsername(username).orElseThrow(() -> {
+    public boolean deleteImage(String username) {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        if (!username.equals(auth.getName())) {
+            throw new ForbiddenUserException();
+        }
+
+        final User user = getUserByUsername(username).orElseThrow(() -> {
             LOGGER.error("Cannot update image for user {} because it does not exist", username);
             return new UserNotFoundException();
         });
 
-        Image image = null;
-        if (user.getImage() != null)
-            image = user.getImage();
+        final Image image = user.getImage();
 
-        user.deleteImage();
-
-        if (image != null) imageService.deleteImage(image);
-
-        return user;
+        if (image != null) {
+            user.deleteImage();
+            imageService.deleteImage(image);
+            return true;
+        }
+        return false;
     }
 
     @Override
@@ -112,6 +121,11 @@ public class UserServiceImpl implements UserService {
     @Async
     @Transactional
     public void deleteUser(String username) {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        if (!username.equals(auth.getName())) {
+            throw new ForbiddenUserException();
+        }
+
         User user = getUserByUsername(username).orElseThrow(() -> {
             LOGGER.error("Cannot delete User {} because it does not exist", username);
             return new UserNotFoundException();
